Make using the UI overlay optional and allow using SafeArea
When `country_list_pick` is used inside an app that uses `SafeArea`,
using the widget's country selection list messes with the notification
area of the device. This commit allows clients to choose whether
`SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle()` should be
applied and allows them to have the country selection list wrapped
inside a `SafeArea` widget.